Address pytorch versioning issues. (#820)
* This commit address version compatibility issues with pytorch.

Many new features of physicsnemo's distributed utilities, targeting domain parallelism,
require pytorch's DTensor package which was introduced in pytorch 2.6.0.  But, we don't
want to limit physicsnemo usage unnecessarily.

This commit introduces version checking utilities, which are then aplied to ShardTensor.
If torch is below 2.6.0, the distributed utilities will not import ShardTensor but
will still work.  If a user attempts to import ShardTensor directly, avoiding the
__init__.py  file, the version checking utilities will raise an exception.

Tests on shard tensor are likewise skipped if torch 2.6.0 is not installed.

Finally, an additional test file is included to validate the version checking tools.

* This commit further protects against older versions of pytorch
- change shard tensor minimum version to 2.5.9 to accomodate alpha release of 2.6.0a
- set minimum pytorch version for DeviceMesh to 2.4.0
- introduce function decorator that raises an exception when unavailable functions are called.
- adds a little more protection in the tests to differntiate,

---------