Add unit test for torch compile (#872)
* Add test for fullgraph compile

Graph breaks in model code are detrimental to performance when
models are compiled. Add a test for fullgraph compile of models.
For models where fullgraph capture works (fno, songunet), this test will
ensure that model code updates dont break it. For the other models
(healpix, graphcast) this test shows graphbreaks so they can be fixed.

Signed-off-by: Ramu Ramamurthy <ramu.ramamurthy@gmail.com>

* add compile mode

Signed-off-by: Ramu Ramamurthy <ramu.ramamurthy@gmail.com>

* make test faster, and add recompile check

Signed-off-by: Ramu Ramamurthy <rramamurthy@nvidia.com>

---------

Signed-off-by: Ramu Ramamurthy <ramu.ramamurthy@gmail.com>
Signed-off-by: Ramu Ramamurthy <rramamurthy@nvidia.com>
Co-authored-by: Ramu Ramamurthy <rramamurthy@nvidia.com>
Co-authored-by: Charlelie Laurent <84199758+CharlelieLrt@users.noreply.github.com>